Refactor to combine Track + TrackLine -> Track
